+++ djpcms/__init__.py
@@ -1,6 +1,6 @@
 '''Dynamic content management system using Javascript and Python'''
 
-VERSION = (0, 7, 2)
+VERSION = (0, 7, 3)
 
 
 def get_version():